+++ CHANGELOG.rst
@@ -21,6 +21,9 @@ Changelog
   * Tracker - issues relation
   * IssueStatus - issues relation
 
+- Added: Introduced a ``values()`` method in a ResourceSet which returns ValuesResourceSet — a
+  ResourceSet subclass that returns dictionaries when used as an iterable, rather than resource-instance
+  objects (see `docs <http://python-redmine.readthedocs.org/operations.html#filter>`__ for details)
 - Added: Introduced ``update()`` and ``delete()`` methods in a ResourceSet object which allow to
   bulk update or bulk delete all resources in a ResourceSet object (see `docs <http://python-redmine.
   readthedocs.org/operations.html#filter>`__ for details)